refactor(RestaurantItem): RestaurantItem을 TS로 마이그레이션
